Отрефакторил MapExtensions
Сгруппировал расширения в отдельные файлы. Также добавил несколько новых расширений для маппинга.
