+++ app/src/main/kotlin/com/designdrivendevelopment/kotelok/repositoryImplementations/getStatisticsRepository/GetStatisticsRepositoryImpl.kt
@@ -4,9 +4,9 @@ import com.designdrivendevelopment.kotelok.entities.TotalDictionaryStat
 import com.designdrivendevelopment.kotelok.entities.TotalStat
 import com.designdrivendevelopment.kotelok.persistence.daos.StatisticsDao
 import com.designdrivendevelopment.kotelok.persistence.queryResults.DictStatQueryResult
-import com.designdrivendevelopment.kotelok.repositoryImplementations.toDictionaryStat
-import com.designdrivendevelopment.kotelok.repositoryImplementations.toTotalDictionaryStat
-import com.designdrivendevelopment.kotelok.repositoryImplementations.toWordDefinitionStat
+import com.designdrivendevelopment.kotelok.repositoryImplementations.extensions.toDictionaryStat
+import com.designdrivendevelopment.kotelok.repositoryImplementations.extensions.toTotalDictionaryStat
+import com.designdrivendevelopment.kotelok.repositoryImplementations.extensions.toWordDefinitionStat
 import com.designdrivendevelopment.kotelok.screens.profile.GetStatisticsRepository
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
