+++ app/src/main/kotlin/com/designdrivendevelopment/kotelok/repositoryImplementations/learnableDefinitionsRepository/CardsLearnableDefinitionsRepository.kt
@@ -2,7 +2,7 @@ package com.designdrivendevelopment.kotelok.repositoryImplementations.learnableD
 
 import com.designdrivendevelopment.kotelok.entities.LearnableDefinition
 import com.designdrivendevelopment.kotelok.persistence.daos.CardsLearnableDefDao
-import com.designdrivendevelopment.kotelok.repositoryImplementations.toLearnableDef
+import com.designdrivendevelopment.kotelok.repositoryImplementations.extensions.toLearnableDef
 import com.designdrivendevelopment.kotelok.screens.trainers.LearnableDefinitionsRepository
 import java.util.Date
 import kotlinx.coroutines.Dispatchers
