+++ app/src/main/kotlin/com/designdrivendevelopment/kotelok/repositoryImplementations/lookupWordDefinitionRepository/LookupWordDefRepositoryImpl.kt
@@ -2,7 +2,7 @@ package com.designdrivendevelopment.kotelok.repositoryImplementations.lookupWord
 
 import com.designdrivendevelopment.kotelok.entities.WordDefinition
 import com.designdrivendevelopment.kotelok.persistence.daos.WordDefinitionsDao
-import com.designdrivendevelopment.kotelok.repositoryImplementations.toWordDefinition
+import com.designdrivendevelopment.kotelok.repositoryImplementations.extensions.toWordDefinition
 import com.designdrivendevelopment.kotelok.screens.dictionaries.lookupWordDefinitionsScreen.LookupWordDefinitionsRepository
 import com.designdrivendevelopment.kotelok.yandexDictApi.YandexDictionaryApiService
 import com.designdrivendevelopment.kotelok.yandexDictApi.responses.YandexDictionaryResponse
